requirements.txt
kaitaistruct==0.9
Pillow==8.3.1
